Update README.md
I designed a Database which the  HR department would  depend on to have full visibility of the company’s employees and their data using SQL. Now instead of maintaining this data on disparate spreadsheets, the company wants to have a single Database where all this tables will be housed to avoid the risk of losing them and to improve reporting efficiency.

The primary objective of the database design is to create a structured and relational data model that efficiently manages employee information, department structures, project management assignments, and geographical data. This database is designed to support various business operations, including employee management, departmental analysis, and project oversight, while enabling robust data retrieval and reporting functionalities. The design ensures data integrity, reduces redundancy, and allows for seamless data aggregation and analysis, which are essential for informed decision-making within the organization.